+++ src/cli/highlighter.rs
@@ -18,7 +18,8 @@ pub(crate) fn build_highlighter(
         numbers,
         uuids,
         quotes,
-        ip_addresses,
+        ipv4,
+        ipv6,
         dates,
         durations,
         paths,
@@ -45,10 +46,10 @@ pub(crate) fn build_highlighter(
         b = b.with_date_time_highlighters(dates);
     }
     if base.contains(Base::Ipv4) {
-        b = b.with_ip_v4_highlighter(ip_addresses.into());
+        b = b.with_ip_v4_highlighter(ipv4);
     }
     if extras.contains(&Extra::Ipv6) {
-        b = b.with_ip_v6_highlighter(ip_addresses.into());
+        b = b.with_ip_v6_highlighter(ipv6);
     }
     if extras.contains(&Extra::JvmStackTrace) {
         b = b.with_jvm_stack_trace_highlighter(jvm_stack_traces);